+++ setup.php
@@ -18,7 +18,7 @@ use GlpiPlugin\Uxcustomizer\Config;
 use GlpiPlugin\Uxcustomizer\Menu;
 use GlpiPlugin\Uxcustomizer\MenuOrder;
 
-define('PLUGIN_UXCUSTOMIZER_VERSION',          '1.4.0');
+define('PLUGIN_UXCUSTOMIZER_VERSION',          '1.5.0');
 define('PLUGIN_UXCUSTOMIZER_MIN_GLPI_VERSION', '11.0.0');
 define('PLUGIN_UXCUSTOMIZER_MAX_GLPI_VERSION', '11.0.99');